feat: update runOnInterval usage across multiple modules to accept options object
