+++ modules/alerts/apps/organizer/package.json
@@ -20,7 +20,8 @@
 		"@tmlmobilidade/go-alerts-pckg-organize": "*",
 		"@tmlmobilidade/interfaces": "*",
 		"@tmlmobilidade/logger": "*",
-		"@tmlmobilidade/timer": "*"
+		"@tmlmobilidade/timer": "*",
+		"@tmlmobilidade/utils": "*"
 	},
 	"devDependencies": {
 		"@tmlmobilidade/tsconfig": "*",
