+++ modules/alerts/apps/organizer/src/index.ts
@@ -4,49 +4,48 @@ import { organizeAlert } from '@tmlmobilidade/go-alerts-pckg-organize';
 import { alerts } from '@tmlmobilidade/interfaces';
 import { Logger } from '@tmlmobilidade/logger';
 import { Timer } from '@tmlmobilidade/timer';
+import { runOnInterval } from '@tmlmobilidade/utils';
 
 /* * */
 
-(async function init() {
-	const runOnInterval = async () => {
-		//
+const main = async () => {
+	//
 
-		Logger.init();
+	Logger.init();
 
-		const globalTimer = new Timer();
+	const globalTimer = new Timer();
 
-		//
-		// Get all Alert documents from the database
+	//
+	// Get all Alert documents from the database
 
-		const allAlertsData = await alerts.findMany({}, { sort: { publish_start_date: -1 } });
+	const allAlertsData = await alerts.findMany({}, { sort: { publish_start_date: -1 } });
 
-		Logger.info(`Found ${allAlertsData.length} alerts.`);
+	Logger.info(`Found ${allAlertsData.length} alerts.`);
 
-		//
-		// Loop through all alerts and request updated attributes for each document
+	//
+	// Loop through all alerts and request updated attributes for each document
 
-		for (const [alertIndex, alertData] of allAlertsData.entries()) {
-			try {
-				//
+	for (const [alertIndex, alertData] of allAlertsData.entries()) {
+		try {
+			//
 
-				Logger.info(`[${allAlertsData.length - alertIndex}/${allAlertsData.length}] Processing Alert ${alertData._id}...`);
+			Logger.info(`[${allAlertsData.length - alertIndex}/${allAlertsData.length}] Processing Alert ${alertData._id}...`);
 
-				const organizedAlertData = await organizeAlert(alertData);
+			const organizedAlertData = await organizeAlert(alertData);
 
-				await alerts.updateById(alertData._id, organizedAlertData);
+			await alerts.updateById(alertData._id, organizedAlertData);
 
-				//
-			}
-			catch (error) {
-				Logger.error(`Error processing Alert ${alertData._id}:`, error);
-			}
+			//
+		} catch (error) {
+			Logger.error(`Error processing Alert ${alertData._id}:`, error);
 		}
+	}
 
-		Logger.terminate(`Organization completed in ${globalTimer.get()}`);
+	Logger.terminate(`Organization completed in ${globalTimer.get()}`);
 
-		//
+	//
+};
 
-		setTimeout(runOnInterval, 300_000); // 5 minutes in milliseconds
-	};
-	runOnInterval();
-})();
+/* * */
+
+await runOnInterval(main, { intervalMs: 300_000 }); // 5 minutes in milliseconds
