+++ modules/controller/apps/rides-acceptor/src/index.ts
@@ -13,7 +13,6 @@ import { Interval } from 'luxon';
 /* * */
 
 const SYNC_DAYS_BACK = 3;
-const RUN_INTERVAL = 600_000; // 10 minutes
 
 async function createRideAcceptances(ride: Ride) {
 	try {
@@ -205,4 +204,4 @@ async function main() {
 
 //
 
-runOnInterval(main, RUN_INTERVAL);
+await runOnInterval(main, { intervalMs: 600_000 }); // Run every 10 minutes
