+++ modules/controller/apps/rides-feeder/src/index.ts
@@ -118,4 +118,4 @@ async function main() {
 
 /* * */
 
-runOnInterval(main, 60_000); // 1 minute
+await runOnInterval(main, { intervalMs: 60_000 }); // 1 minute
