+++ modules/controller/apps/sams-feeder/src/index.ts
@@ -194,4 +194,4 @@ async function main() {
 
 /* * */
 
-runOnInterval(main, 43_200_000); // 12 hours
+await runOnInterval(main, { intervalMs: 43_200_000 }); // 12 hours
