+++ modules/exporter/apps/export-gtfs-merged/src/index.ts
@@ -5,4 +5,4 @@ import { runOnInterval } from '@tmlmobilidade/utils';
 
 /* * */
 
-await runOnInterval(main, 60_000); // 1 minute
+await runOnInterval(main, { intervalMs: 60_000 }); // 1 minute
