+++ modules/performance/apps/sync-metrics-daily/src/index.ts
@@ -38,4 +38,4 @@ async function main() {
 	//
 }
 
-runOnInterval(main, 86_400_000); // 1 day
+await runOnInterval(main, { intervalMs: 86_400_000 }); // 1 day
