+++ modules/replicator/apps/sync-apex-on-board-sales/package.json
@@ -17,6 +17,7 @@
 		"lint:fix": "eslint . --fix"
 	},
 	"dependencies": {
+		"@tmlmobilidade/dates": "*",
 		"@tmlmobilidade/go-replicator-pckg-parse": "*",
 		"@tmlmobilidade/go-replicator-pckg-sync": "*",
 		"@tmlmobilidade/interfaces": "*",
