+++ modules/stops/apps/organizer/src/index.ts
@@ -42,4 +42,4 @@ async function main() {
 	//
 }
 
-runOnInterval(main, 300_000); // 5 minutes
+await runOnInterval(main, { intervalMs: 300_000 }); // 5 minutes
